refactor: move crud prose tests into the right spot
